backend/pyproject.toml
[tool.mypy]
mypy_path = ["backend/recipes"]
plugins = ["mypy_django_plugin.main"]


[tool.django-stubs]
django_settings_module = "recipes.settings"

[tool.poetry]
name = "recipe-backend"
version = "0.1.0"
description = "DRF for recipe backend"
authors = ["Akxary <pollxx97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.1.4"
django-enum = "2.0.2"
gunicorn = "23.0.0"
psycopg2-binary = "2.9.10"
djangorestframework = "3.15.2"
drf-spectacular = "0.28.0"
typing-extensions = "4.12.2"
redis = "^5.2.1"
pyjwt = "^2.10.1"
celery = "^5.4.0"


[tool.poetry.group.dev.dependencies]
black = "24.10.0"
pytest = "8.3.4"
pytest-django = "4.9.0"
django-stubs = "5.1.1"
django-stubs-ext = "5.1.1"
djangorestframework-stubs = "3.15.2"
types-redis = "^4.6.0.20241004"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
